requirements.txt
python-dotenv==1.0.0
boto3==1.34.0
requests==2.31.0
pytest==7.4.0
twilio==8.10.0
